+++ docs/changelog.rst
@@ -22,11 +22,13 @@ v0.15.0 (UNRELEASED)
   both functions and properties have been moved into Tracklist to have more
   modularity:
       - Properties converted into functions that need arguments:
-        :meth:`get_tl_track_at_eot`, :meth:`get_tl_track_at_next`, 
-        :meth:`get_tl_track_at_previous` and :meth:`tracklist_position`
+        :meth:`tracklist_position`
       - Properties moved: :attr:`random`, :attr:`repeat`, :attr:`consume` and
         :attr:`single`
-      - Method created: :meth:`mark`
+      - Method created from properties: :meth:`next_track` from 
+        `tl_track_at_next`, :meth:`eot_track` from ´tl_track_at_eot` and
+        :meth:`previous_track` from `tl_track_at_previous`
+      - Method created to separe functionality: :meth:`mark` 
 - Tracklist's get_tl_track_at_* and tracklist_position now need a tl_track as a
   reference to give an answer.